# chore(fs): speed up case normalization